+++ src/oakutils/_api_camera.py
@@ -30,6 +30,7 @@ from typing import TYPE_CHECKING, Callable, Iterable
 import depthai as dai
 from typing_extensions import TypeAlias
 
+from ._basic import create_device
 from .calibration import CalibrationData, ColorCalibrationData, get_camera_calibration
 from .tools.display import DisplayManager, get_smaller_size
 
@@ -239,11 +240,7 @@ class ApiCamera:
         with self._start_condition:
             self._start_condition.wait()
 
-        if self._mxid is not None:
-            device_info: dai.DeviceInfo = dai.DeviceInfo(self._mxid)
-            device_object = dai.Device(self._pipeline, device_info)
-        else:
-            device_object = dai.Device(self._pipeline)
+        device_object = create_device(self._pipeline, device_id=self._mxid)
         with device_object as device:
             # run any custom devices calls added ahead of time
             for custom in self._custom_device_calls: