+++ src/oakutils/vpu/_vpu.py
@@ -17,12 +17,17 @@ from __future__ import annotations
 import atexit
 import logging
 from pathlib import Path
+from queue import Empty, Queue
 from threading import Condition, Thread
 from typing import TYPE_CHECKING
 
 import depthai as dai
+import numpy as np
 
+from oakutils import create_device
 from oakutils.nodes import (
+    MobilenetData,
+    YolomodelData,
     create_mobilenet_detection_network,
     create_neural_network,
     create_xin,
@@ -31,10 +36,7 @@ from oakutils.nodes import (
 )
 
 if TYPE_CHECKING:
-    import numpy as np
     from typing_extensions import Self
-
-    from oakutils.nodes import MobilenetData, YolomodelData
 
 _log = logging.getLogger(__name__)
 
@@ -57,20 +59,26 @@ class VPU:
             Examples: "14442C108144F1D000", "192.168.1.44", "3.3.3"
 
         """
-        self._blob_path: str | None = None
-        self._pipeline: dai.Pipeline | None = None
-        self._xin: dai.node.XLinkIn | list[dai.node.XLinkIn] | None = None
-        self._input_names: list[str] | None = None
-        self._nn: dai.node.NeuralNetwork | None = None
-        self._xout: dai.node.XLinkOut | None = None
+        # general attributes
+        self._mxid: str | None = device_id
+        self._pipeline: dai.Pipeline = dai.Pipeline()
         self._thread: Thread | None = None
         self._start_condition = Condition()
-        self._condition = Condition()
+        self._data_queue: Queue[list[np.ndarray | list[np.ndarray]]] = Queue()
+        self._result_queue: Queue[list[np.ndarray | dai.ImgDetections]] = Queue()
         self._stopped = False
-        self._data: np.ndarray | list[np.ndarray] | None = None
-        self._result: np.ndarray | None = None
 
-        self._mxid: str | None = device_id
+        # attributes for multi model execution
+        self._blob_paths: list[str] = []
+        self._inputnames: list[list[str] | None] = []
+        self._nns: list[dai.node.NeuralNetwork] = []
+        self._xins: list[dai.node.XLinkIn | list[dai.node.XLinkIn]] = []
+        self._xouts: list[dai.node.XLinkOut] = []
+        self._xin_names: list[str | list[str]] = []
+        self._xout_names: list[str] = []
+
+        # mode checking
+        self._multimode: bool = False
 
         atexit.register(self.stop)
 
@@ -80,8 +88,10 @@ class VPU:
 
     def __call__(
         self: Self,
-        data: np.ndarray | list[np.ndarray],
-    ) -> np.ndarray | dai.ImgDetections:
+        data: np.ndarray | list[np.ndarray] | list[np.ndarray | list[np.ndarray]],
+        *,
+        safe: bool | None = None,
+    ) -> np.ndarray | dai.ImgDetections | list[np.ndarray | dai.ImgDetections]:
         """
         Use to run an inference on the VPU.
 
@@ -89,11 +99,14 @@ class VPU:
         ----------
         data : np.ndarray | list[np.ndarray]
             The data to run an inference on.
-
+        safe : bool, optional
+            If True, will evaluate the data before sending to the VPU.
+            If False, will send data directly to the VPU.
+            By default None, which will use safe mode.
 
         Returns
         -------
-        np.ndarray | dai.ImgDetections
+        np.ndarray | dai.ImgDetections | list[np.ndarray | dai.ImgDetections]
             The result of the inference.
 
         Raises
@@ -102,189 +115,328 @@ class VPU:
             If the blob path is not set.
 
         """
-        return self.run(data)
+        return self.run(data, safe=safe)
 
     def stop(self: Self) -> None:
         """Use to stop the VPU."""
         self._stopped = True
-        with self._condition:
-            self._condition.notify()
         if self._thread is not None:
             if self._thread.is_alive():
                 self._thread.join()
             else:
                 pass
 
+    def _reset(self: Self) -> None:
+        """Reset the VPU attributes. Should only be called in reconfiguration."""
+        # stop the thread if active
+        if self._thread is not None:
+            self.stop()
+        self._stopped = False
+        # reset list attributes
+        self._blob_paths = []
+        self._xins = []
+        self._inputnames = []
+        self._nns = []
+        self._xouts = []
+        self._xin_names = []
+        self._xout_names = []
+        # reset pipeline
+        self._pipeline = dai.Pipeline()
+
     def reconfigure(
         self: Self,
         blob_path: str | Path,
         input_names: list[str] | None = None,
-        yolo_data: YolomodelData | None = None,
-        mobilenet_data: MobilenetData | None = None,
-        *,
-        is_yolo_model: bool | None = None,
-        is_mobilenet_model: bool | None = None,
+        model_data: YolomodelData | MobilenetData | None = None,
     ) -> None:
         """
-        Use to reconfigure the VPU with a new blob file.
+        Use to reconfigure the VPU with a single new blob file.
 
         Parameters
         ----------
         blob_path : str | Path
             The path to the blob file.
-        input_names : list[str]
+        input_names : list[str], optional
             The names of the input layers. Defaults to None.
-        is_yolo_model : bool
-            Whether or not the blob file is a YOLO model. Defaults to None.
-        yolo_data : YolomodelData
-            The YOLO model data. Defaults to None.
-            If is_yolo_model is True, must not be None
-        is_mobilenet_model : bool
-            Whether or not the blob file is a Mobilenet model. Defaults to None.
-        mobilenet_data : MobilenetData
-            The Mobilenet model data. Defaults to None.
-            If is_mobilenet_model is True, must not be None
+        model_data : YolomodelData | MobilenetData, optional
+            The model data. Defaults to None.
+            Can be used to set the YoloModelData or MobilenetData.
+            If None, then a generic neural network will be created.
 
         Raises
         ------
-        ValueError
-            If is_yolo_model is True and yolo_data is None
-            If is_mobilenet_model is True and mobilenet_data is None
-            If is_yolo_model is True and is_mobilenet_model is True
         FileNotFoundError
             If the blob file does not exist.
+        TypeError
+            If model_data is not YoloModelData or MobilenetData.
 
         """
-        if is_yolo_model is None:
-            is_yolo_model = False
-        if is_mobilenet_model is None:
-            is_mobilenet_model = False
-        if is_yolo_model and is_mobilenet_model:
-            err_msg = "Cannot be both YOLO model and Mobilenet model."
-            raise ValueError(err_msg)
-        # stop the VPU if it is running
-        if self._thread is not None:
-            self.stop()
-        self._stopped = False
-        self._blob_path = (
-            blob_path if isinstance(blob_path, str) else str(blob_path.resolve())
+        self._multimode = False
+        # repackage as input to backend function
+        self._reconfigure(
+            [blob_path],
+            [input_names],
+            [model_data],
         )
-        if not Path.exists(Path(self._blob_path)):
-            err_msg = f"Blob file {self._blob_path} does not exist."
-            raise FileNotFoundError(err_msg)
-        # create pipeline with neural network
-        self._pipeline = dai.Pipeline()
-        if input_names is None:
-            self._xin = create_xin(self._pipeline, "vpu_in")
-            if not is_yolo_model and not is_mobilenet_model:
-                _log.debug("Reconfiguring VPU with single input.")
-                self._nn = create_neural_network(
-                    self._pipeline,
-                    self._xin.out,
-                    Path(self._blob_path),
-                )
-                self._xout = create_xout(self._pipeline, self._nn.out, "vpu_out")
-            if is_yolo_model:
-                if yolo_data is None:
-                    err_msg = "YOLO data must not be None."
-                    raise ValueError(err_msg)
-                _log.debug("Reconfiguring VPU with YOLO model.")
-                self._nn = create_yolo_detection_network(
-                    self._pipeline,
-                    self._xin.out,
-                    Path(self._blob_path),
-                    yolo_data=yolo_data,
-                )
-                if self._nn is None:
-                    err_msg = "Neural network is None, major internal error occured."
-                    raise RuntimeError(
-                        err_msg,
-                    )
-                self._xout = create_xout(self._pipeline, self._nn.out, "vpu_out")
-            if is_mobilenet_model:
-                if mobilenet_data is None:
-                    err_msg = "Mobilenet data must not be None."
-                    raise ValueError(err_msg)
-                _log.debug("Reconfiguring VPU with Mobilenet model.")
-                self._nn = create_mobilenet_detection_network(
-                    self._pipeline,
-                    self._xin.out,
-                    Path(self._blob_path),
-                    mobilenet_data=mobilenet_data,
-                )
-                if self._nn is None:
-                    err_msg = "Neural network is None, major internal error occured."
-                    raise RuntimeError(
-                        err_msg,
-                    )
-                self._xout = create_xout(self._pipeline, self._nn.out, "vpu_out")
+
+    def reconfigure_multi(
+        self: Self,
+        blob_paths: list[str | Path],
+        input_names: list[list[str] | None] | None = None,
+        modeldata: list[YolomodelData | MobilenetData | None] | None = None,
+    ) -> None:
+        """
+        Reconfigure the VPU with multiple blob files.
+
+        Parameters
+        ----------
+        blob_paths : list[str | Path]
+            The paths to the blob files.
+        input_names : list[list[str] | None]
+            The names of the input layers. Defaults to None.
+            Should be filled in if a model has multiple inputs.
+        modeldata : list[YolomodelData | MobilenetData | None]
+            The model data. Defaults to None.
+            Should be filled in if the model is a YOLO or Mobilenet model.
+            If None, then a generic neural network will be created.
+
+        Raises
+        ------
+        FileNotFoundError
+            If a blob file does not exist.
+        ValueError
+            If input_names is not None and does not match the length of blob_paths.
+            If modeldata is not None and does not match the length of blob_paths.
+        TypeError
+            If modeldata is not YoloModelData or MobilenetData.
+
+        """
+        self._multimode = True
+        self._reconfigure(
+            blob_paths,
+            input_names,
+            modeldata,
+        )
+
+    def _reconfigure(
+        self: Self,
+        blob_paths: list[str | Path],
+        input_names: list[list[str] | None] | None = None,
+        modeldata: list[YolomodelData | MobilenetData | None] | None = None,
+    ) -> None:
+        """Handle reconfiguration of the VPU."""
+        # reset the VPU
+        self._reset()
+
+        self._blob_paths = [
+            blob_path if isinstance(blob_path, str) else str(blob_path.resolve())
+            for blob_path in blob_paths
+        ]
+        # validate the blob files
+        for idx, blob_path in enumerate(self._blob_paths):
+            if not Path.exists(Path(blob_path)):
+                err_msg = f"Blob file #{idx}: {blob_path}, does not exist."
+                raise FileNotFoundError(err_msg)
+
+        # validate the input_names
+        if input_names is not None:
+            if len(input_names) != len(self._blob_paths):
+                err_msg = "Input names must match the number of blob files."
+                raise ValueError(err_msg)
         else:
-            _log.debug("Reconfiguring VPU with multiple inputs.")
-            self._input_names = input_names
-            self._xin = []
-            for name in self._input_names:
-                self._xin.append(create_xin(self._pipeline, name))
-            self._nn = create_neural_network(
-                self._pipeline,
-                [xin.out for xin in self._xin],
-                Path(self._blob_path),
-                self._input_names,
+            input_names = [None] * len(self._blob_paths)
+        self._inputnames = input_names
+
+        # validate the modeldata
+        if modeldata is not None:
+            if len(modeldata) != len(self._blob_paths):
+                err_msg = "Model data must match the number of blob files."
+                raise ValueError(err_msg)
+        else:
+            modeldata = [None] * len(self._blob_paths)
+
+        # allocate each network in the pipeline
+        for idx, (bpath, iname, mdata) in enumerate(
+            zip(self._blob_paths, self._inputnames, modeldata),
+        ):
+            # allocate input XLink nodes
+            if iname is None:
+                xin_name = f"vpu_in_{idx}"
+                self._xin_names.append(xin_name)
+                self._xins.append(create_xin(self._pipeline, xin_name))
+            else:
+                m_xins = []
+                m_xin_names = []
+                for name in iname:
+                    xin_name = f"vpu_in_{idx}_{name}"
+                    m_xin_names.append(xin_name)
+                    m_xins.append(create_xin(self._pipeline, name))
+                self._xin_names.append(m_xin_names)
+                self._xins.append(m_xins)
+
+            # allocate neural network
+            if mdata is None:
+                # handle single or multi link
+                if isinstance(self._xins[-1], list):
+                    links = [xin.out for xin in self._xins[-1]]
+                else:
+                    links = self._xins[-1].out  # type: ignore[assignment]
+                self._nns.append(
+                    create_neural_network(
+                        self._pipeline,
+                        links,
+                        Path(bpath),
+                        iname,
+                    ),
+                )
+            elif isinstance(mdata, YolomodelData):
+                if isinstance(self._xins[-1], list):
+                    err_msg = "Yolo model data must have a single input."
+                    raise ValueError(err_msg)
+                self._nns.append(
+                    create_yolo_detection_network(
+                        self._pipeline,
+                        self._xins[-1].out,
+                        Path(bpath),
+                        yolo_data=mdata,
+                    ),
+                )
+            elif isinstance(mdata, MobilenetData):
+                if isinstance(self._xins[-1], list):
+                    err_msg = "Mobilenet model data must have a single input."
+                    raise ValueError(err_msg)
+                self._nns.append(
+                    create_mobilenet_detection_network(
+                        self._pipeline,
+                        self._xins[-1].out,
+                        Path(bpath),
+                        mobilenet_data=mdata,
+                    ),
+                )
+            else:
+                err_msg = "Model data must be YoloModelData or MobilenetData."
+                raise TypeError(err_msg)
+
+            # allocate output XLink nodes
+            xout_name = f"vpu_out_{idx}"
+            self._xout_names.append(xout_name)
+            self._xouts.append(
+                create_xout(self._pipeline, self._nns[-1].out, xout_name),
             )
-            self._xout = create_xout(self._pipeline, self._nn.out, "vpu_out")
-        # reallocate the device thread
-        self._thread = Thread(target=self._run, daemon=True)
+
+        # create the thread
+        self._thread = Thread(target=self._run_thread, daemon=True)
         self._thread.start()
+
+        # wait for thread to start
         with self._start_condition:
             self._start_condition.wait()
 
-    def _run(self: Self) -> None:
-        """Use in a thread to process the data on the VPU."""
+    def _run_thread(
+        self: Self,
+    ) -> None:
+        """
+        Use in a thread to process neural networks on the VPU.
+
+        Handles multi-input networks, and an arbitrary number of networks.
+        Will pre-allocate all buffers and queues for each network.
+
+        Raises
+        ------
+        RuntimeError
+            If the pipeline is not set.
+            If the data does not match queues and buffers.
+
+        """
+        # check pipeline, should always be set
         if self._pipeline is None:
             err_msg = "Pipeline not set."
             raise RuntimeError(err_msg)
-        if self._mxid is not None:
-            device_info: dai.DeviceInfo = dai.DeviceInfo(self._mxid)
-            device_object = dai.Device(self._pipeline, device_info)
-        else:
-            device_object = dai.Device(self._pipeline)
+        # create the device
+        device_object = create_device(self._pipeline, device_id=self._mxid)
         with device_object as device:
-            _log.debug("VPU thread started.")
+            # pre-fetch queues and allocate buffers
+            input_buffers: list[dai.Buffer | list[dai.Buffer]] = []
+            input_queues: list[dai.DataInputQueue | list[dai.DataInputQueue]] = []
+            output_queues: list[dai.DataOutputQueue] = []
+            for xin_name, xout_name in zip(self._xin_names, self._xout_names):
+                # handle multi-input networks
+                if isinstance(xin_name, list):
+                    in_buffers = []
+                    in_queues = []
+                    for xname in xin_name:
+                        in_buffers.append(dai.Buffer())
+                        in_queues.append(device.getInputQueue(xname))  # type: ignore[attr-defined]
+                    input_buffers.append(in_buffers)
+                    input_queues.append(in_queues)
+                else:
+                    input_buffers.append(dai.Buffer())
+                    input_queues.append(device.getInputQueue(xin_name))  # type: ignore[attr-defined]
+                # all networks have one output
+                output_queues.append(device.getOutputQueue(xout_name))  # type: ignore[attr-defined]
+
+            # notify the main thread that VPU is ready
+            # this will allow the reconfigure call to return
             with self._start_condition:
                 self._start_condition.notify()
+
+            # loop until stopped
             while not self._stopped:
-                with self._condition:
-                    self._condition.wait()
-                _log.debug("VPU thread notified.")
                 if self._stopped:
-                    _log.debug("VPU stopped, breaking.")
                     break
-                if isinstance(self._xin, list):
-                    _log.debug("Sending multi-value data to VPU.")
-                    if self._input_names is None:
-                        err_msg = "Input names not set."
-                        raise RuntimeError(err_msg)
-                    if self._data is None or not isinstance(self._data, list):
-                        err_msg = "Data not set or data is not a list."
-                        raise RuntimeError(err_msg)
-                    for name, data in zip(self._input_names, self._data):
-                        buff = dai.Buffer()
-                        buff.setData(data)
-                        device.getInputQueue(name).send(buff)  # type: ignore[attr-defined]
-                else:
-                    _log.debug("Sending single-value data to VPU.")
-                    buff = dai.Buffer()
-                    # setData takes list[int] or ndarray, mypy cannot handle this
-                    buff.setData(self._data)  # type: ignore[arg-type]
-                    device.getInputQueue("vpu_in").send(buff)  # type: ignore[attr-defined]
-                self._result = device.getOutputQueue("vpu_out").get()  # type: ignore[attr-defined]
-                _log.debug("VPU result received, notifying primary thread.")
-                with self._condition:
-                    self._condition.notify()
+
+                # get data
+                try:
+                    all_data: list[np.ndarray | list[np.ndarray]] = (
+                        self._data_queue.get(timeout=0.1)
+                    )
+                except Empty:
+                    continue
+
+                # push data to networks
+                for idx, (data, buffers, queues) in enumerate(
+                    zip(all_data, input_buffers, input_queues),
+                ):
+                    # handle multi-input networks
+                    if (
+                        isinstance(buffers, list)
+                        or isinstance(queues, list)
+                        or isinstance(data, list)
+                    ):
+                        # check for mismatched types
+                        if (
+                            isinstance(buffers, list)
+                            and isinstance(queues, list)
+                            and isinstance(data, list)
+                        ):
+                            for sub_data, buff, queue in zip(data, buffers, queues):
+                                buff.setData(sub_data)
+                                queue.send(buff)
+                        else:
+                            err_msg = f"Error Network #{idx}: Mismatched data, buffers, and queues."
+                            err_msg += (
+                                " All must be lists or all must be single objects."
+                            )
+                            err_msg += " Mismatched types: "
+                            err_msg += f"buffers: {type(buffers)}, queues: {type(queues)}, data: {type(data)}"
+                            raise TypeError(err_msg)
+                    # all yolo/mobilenet and most networks will be single input
+                    else:
+                        buffers.setData(data)
+                        queues.send(buffers)
+
+                # get the results
+                results: list[np.ndarray | dai.ImgDetections] = [
+                    queue.get() for queue in output_queues  # type: ignore[misc]
+                ]
+                self._result_queue.put(results)
 
     def run(
         self: Self,
-        data: np.ndarray | list[np.ndarray],
-    ) -> np.ndarray | dai.ImgDetections:
+        data: np.ndarray | list[np.ndarray] | list[np.ndarray | list[np.ndarray]],
+        *,
+        safe: bool | None = None,
+    ) -> np.ndarray | dai.ImgDetections | list[np.ndarray | dai.ImgDetections]:
         """
         Use to run an inference on the VPU.
 
@@ -292,24 +444,25 @@ class VPU:
         ----------
         data : np.ndarray | list[np.ndarray]
             The data to run an inference on.
-
+        safe : bool, optional
+            If True, will evaluate the data before sending to the VPU.
+            If False, will send data directly to the VPU.
+            By default None, which will use safe mode.
 
         Returns
         -------
-        np.ndarray | dai.ImgDetections
+        np.ndarray | dai.ImgDetections | list[np.ndarray | dai.ImgDetections]
             The result of the inference.
 
         Raises
         ------
         RuntimeError
-            If the blob path is not set.
-            If the VPU thread is not set or alive. (Should not happen.)
+            If the VPU thread is not set or alive. Will occur if not configured.
             If the VPU result is None.
 
         """
-        if self._blob_path is None:
-            err_msg = "Blob path not set."
-            raise RuntimeError(err_msg)
+        if safe is None:
+            safe = True
         if self._thread is None:
             err_msg = "VPU thread not set."
             raise RuntimeError(err_msg)
@@ -317,13 +470,52 @@ class VPU:
             err_msg = "VPU thread is not alive."
             raise RuntimeError(err_msg)
         _log.debug("VPU run called.")
-        self._data = data
-        with self._condition:
-            self._condition.notify()
-        _log.debug("Notified VPU thread, waiting for result.")
-        with self._condition:
-            self._condition.wait()
-        if self._result is None:
-            err_msg = "VPU result is None."
-            raise RuntimeError(err_msg)
-        return self._result
+        if not self._multimode:
+            result = self._run_single(data, safe=safe)  # type: ignore[arg-type, assignment]
+        else:
+            result = self._run_multi(data, safe=safe)  # type: ignore[arg-type, assignment]
+        return result
+
+    def _run_single(
+        self: Self,
+        data: np.ndarray | list[np.ndarray],
+        *,
+        safe: bool | None = None,
+    ) -> np.ndarray | dai.ImgDetections:
+        if safe is None:
+            safe = True
+        if safe:
+            # evaluate data when it is a list
+            if isinstance(data, list):
+                # in single mode must only contain np.array
+                if any(isinstance(d, list) for d in data):
+                    err_msg = "Configuration with single network expects np.array or list of np.arrays,"
+                    err_msg += " not a list with lists inside."
+                    raise TypeError(err_msg)
+                if not all(isinstance(d, np.ndarray) for d in data):
+                    err_msg = "Configuration with single network expects np.array or list of np.arrays."
+                    raise TypeError(err_msg)
+            self._data_queue.put([data])
+        else:
+            # just send it
+            self._data_queue.put([data])
+        _log.debug("Waiting on VPU result.")
+        return self._result_queue.get()[0]
+
+    def _run_multi(
+        self: Self,
+        data: list[np.ndarray | list[np.ndarray]],
+        *,
+        safe: bool | None = None,
+    ) -> list[np.ndarray | dai.ImgDetections]:
+        if safe is None:
+            safe = True
+        if safe:
+            if isinstance(data, np.ndarray) or not isinstance(data, list):
+                err_msg = "Configuration with multiple networks expects type: list[np.ndarray | list[np.ndarray]]."
+                raise TypeError(err_msg)
+            self._data_queue.put(data)
+        else:
+            self._data_queue.put(data)
+        _log.debug("Waiting on VPU result.")
+        return self._result_queue.get()